backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Validation & Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Client (for external APIs)
httpx>=0.26.0
aiohttp>=3.9.0

# PDF Generation
reportlab>=4.0.0
pypdf>=3.17.0

# AI Integration
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
